Streamlit/requirements.txt
streamlit
pandas
snowflake-connector-python
plotly.express
datetime
pandasql